Do not dispose of root certificate when it is used.
This defers disposal of the root certificate when it might be needed for OCSP staple fetching.
Before this change, the root certificate would get disposed, giving it a null handle. We would then
use this null handle when attempting to build an OCSP request, which OpenSSL would null-deref.

For platforms that don't need the root certificate, they dispose of it.