JIT: fix some asserts in profile incorporation and reconstruction (#82196)
If we're instrumenting a Tier0 method with partial compilation enabled, we may
see divergence between the instrumentation plan phase (which runs before
importation) and the instrumentation phase (which runs after), because the
importer may have altered flow at partial compilation points. Tolerate this.

If we're instrumenting a Tier1 OSR method whose first block is a self-loop,
we may not see a pseudo-edge from the OSR entry to the method's first block
in the profile reconstruction model, because they are the same block.
Tolerate this.

Fixes #82128.