Intrinsically recognize Unsafe APIs in the scanner
Size optimization. If we're in shared code, call to an Unsafe intrinsic would bring generic lookup for the generic dictionary of the generic method. We would not be able to get rid of it later because it would be referenced from the other generic dictionary (the purpose of the scanning phase is to figure out how generic dictionaries will look like when we start code generation - this would be an unused slot).

To add insult to injury, the dictionary would be empty most of the time because these generic methods don't actually do anything with their T.

RyuJIT recognizes these intrinsically and it's not going to ask for the generic dictionary.

Loose generic dictionaries actually add quite a bit of costs because we need to keep track of them in hash tables in the generated executable (type loader might need it, reflection might need them, etc.)

Saves almost 1.1% on BasicMinimalApi (!!!).