Fold canonical templates into canonically equivalent types
Saves 0.6% for BasicMinimalApi.

The type loader in NativeAOT is a "template type loader" - to load new types, it starts with an existing `MethodTable` and makes a copy of it doing adjustments as necessary.

In .NET Native, we used real `MethodTable`s such as `List<object>`. This had limitations because we sometimes didn't know how a real instantiation should look like - at some point we introduced `CanonALike` templates (`List<CanonAlike>`), mostly due to tooling limitations in rhbind.

In NativeAOT, we use canonical types (`List<__Canon>`). But this is potentially a waste of space because we might be able to use a real type if one exists. This switches us to a "best of both worlds" approach - we still model templates as canonical instantiations, but if at object emission time we find out there's a suitable concrete instantiations, we use that.