Update usage of AllowEmptyTelemetry based on changes to the task in the SDK
In https://github.com/dotnet/sdk/pull/30269 the API of the AllowEmptyTelemetry task changed to allow for more granular hashing of the collected telemetry properties. This change was against a servicing branch and flowed into SDK main. Sometime after that change, these targets moved from SDK to Runtime, and so the change was lost.

This re-applies the change to Runtime main to address breaks seen in source-build.