JIT: Assume allocations succeed
We call a NOMEM function when allocations fail that is already marked as
no-return, but at least MSVC is not able to pick up on this fact. The
result is that all operator new calls end up with an unnecessary null
check. This adds an __assume to allow MSVC to recognize that allocations
never fail. For other compilers, it uses the preexisting UNREACHABLE()
macro, but MSVC was not able to recognize this either.

There was also a workaround for an old VSW bug that I've removed.