JIT: Remove JTRUE(relop) invariant in the backend (#82766)
Today the backend (and much of the JIT) allows only JTRUE with a relop
operand, and it is always assumed that this relop appears right before
JTRUE in linear order. This change adds support for JTRUE nodes with
arbitrary integral operands from lowering and onwards. The operand is
also allowed to appear at arbitrary locations in LIR.

With this, we can now enable an optimization that turns
EQ/NE(relop/SETCC, 0) into just a (potentially) reversed condition. This
improves codegen for some cases with SELECTCC and gives us a simple way
to allow for compare chains that feed conditional branches in the
future.

In addition change DISPTREERANGE to handle nodes that consume flags
by also finding the flags definitions as part of looking for data flow.
This is best-effort only, so there is an assert that it is only used in debug
builds.