Update SourceBuild image to be our Centos 8 Stream image.
It's close enough to a plain RHEL8 install that it will serve our purposes for validating our source-build min requirements. It technically can go ahead of RHEL8 (as the Stream distros are ahread releases, not behind ones like CentOS used to be), but I don't think that's enough of a concern as we don't have a good base distro to use otherwise as we can't use RHEL8 directly due to licensing rules.