Use IAttributeOperation for our attribute-based analyzers
Use an operation action instead of a syntax action now that we have an operation action on attributes. The operation action has a much more consistent view of the world (and can go between syntax and structured data much better than AttributeData + AttributeSyntax).

This includes #83238 as the version bump is required to get access to IAttributeOperation.