Remove the second copy of ILLink.Shared project and source code
I compared the two, the one in illink had a couple of small fixes on top of NativeAOT. The only change exclusive to NativeAOT was making the FlowAnnotations public, so I ported that over.

I chose to keep the one in illink. We can discuss moving it around in additional PR.

The rest of this change is solution and project file cleanup.