Shortcut some generic lookups
If the thing we want to look up is the same thing that defines the current context, skip the lookup and use the generic context as-is. For example:

```csharp
class Gen<T>
{
    static void Method1(...) => Method2(...);
    static void Method2(...) => ...;
}
```

Since Method1 is calling Method2, we need to pass the generic context so that we know what the T is. We could either generate it as a dictionary lookup for TypeHandle of Gen<!0>, or we can realize that the current generic context is already the TypeHandle of Gen<!0> and we can use it as-is.