JIT: Generalize handling of commas in block morphing
Eliminate commas early in block morphing, before the rest of the phase
needs to look at it. Do this by extracting their side effects and adding
them on top of the returned result instead. This allows us to handle
arbitrary COMMAs on the source and destination operands in a general
manner.