JIT: stop copying EH tab info into inlinee compiler
Having this info show up for inlinees just confuses things -- if
the inlinee looks in this table and finds a block reference, it will
be to a block that doesn't exist in the inlinee's flow graph.

This has been tripping up profile synthesis, which wants to attribute
some flow to EH entries; inlinees don't have any, but currently look
like they do.

There is one place in inlinee importation where we might ask about EH
properties -- inline pinvoke checks. We're already careful to redirect
this query to the call site block, so in the correspnonding lookup
we must make sure to answer this via the root compiler's EH tab.

If we ever decide to support inlining of methods with EH we'll have
to reconcile all this differently.