Managed testing overhaul
* The managed embedding tests in `UnityEmbedHost.Tests` will now run on our coreclr when ran from our CI script.

* Add `test.cmd` and `test.sh` scripts to make running just the `UnityEmbedHost.Tests` from the command line easier.

* When running the managed embedding tests from the CI scripts, our GC will be used.  (I'm pretty sure)

    When I set `DOTNET_GCName` and ran `dotnet test` on the solution, CoreCLR crashed before ever running tests.  However, when I deferred setting `DOTNET_GCName` until the test process itself, there was no crash.

    I used process explorer to verify the env var was set in the test process.

    Deferring is done by using a run settings file to specify env vars for the test process.

* There are now C# unit tests that will test the native embedding api.

* PInvokes into the native embedding apis are source generated.

* Generate wrappers around the pinvokes to make them easier to use.

* `CoreCLRHostTestingWrappers` is gone.  This is now source generated.  It's essentially the same as the above so it was easy to get both with more or less the same source generation logic.

* Tests that call into our native embedding api are ignored when running from Rider.

Loose Ends

I haven't figured out how to get Rider to use our coreclr to run tests.  Because of that, the test strategy is a bit unresolved.  Currently, I've dulpicated all of the tests.

If I can get Rider (and VS) to run tests on our coreclr, I think maybe we can drop the tests that directly call the managed callback methods.

If I can't get Ride (and VS) to nicely run tests on our coreclr, then the callback tests may need to stay so that there is at least something to run from the IDE.  If this happens, maybe I can somehow come up with some abstract that allows us to reuse 1 set of tests for both purposes.

Other changes & Notes

* Formatting in csproj.  Rider kept wanting to change it.

* There was some refactoring to the source generator to clean things up as I added more to it.

* There was some refactoring to BuildDriver to clean things up as I added mroe to it.